# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Miaota Industrial Agent API
  description: |
    工业智能监控与诊断系统API
    
    ## 功能模块
    - **数据采集**: PLC数据采集与管理
    - **设备管理**: 设备状态监控与配置
    - **告警管理**: 告警规则与通知
    - **数据分析**: 异常检测与趋势分析
    - **知识库**: 故障诊断与知识检索
    - **报表**: 数据导出与可视化
    
  version: v1.0.0-beta1
  contact:
    name: Miaota Team
    email: team@miaota.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: 本地开发环境
  - url: https://api.miaota.local/api/v1
    description: 生产环境
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /devices:
    get:
      tags: [设备管理]
      summary: 获取设备列表
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [s7, modbus]
        - name: status
          in: query
          schema:
            type: string
            enum: [online, offline, error]
      responses:
        '200':
          description: 设备列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'

    post:
      tags: [设备管理]
      summary: 创建设备
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreateRequest'
      responses:
        '201':
          description: 创建成功
  /devices/{device_id}:
    get:
      tags: [设备管理]
      summary: 获取设备详情
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 设备详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [设备管理]
      summary: 更新设备
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdateRequest'
      responses:
        '200':
          description: 更新成功

    delete:
      tags: [设备管理]
      summary: 删除设备
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 删除成功
  /devices/{device_id}/tags:
    get:
      tags: [设备管理]
      summary: 获取设备点位
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 点位列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceTag'

  # ============ 告警 ============
...
  /knowledge/search:
    post:
      tags: [知识库]
      summary: 搜索知识库
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                limit:
                  type: integer
                  default: 5
      responses:
        '200':
          description: 搜索结果
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KnowledgeDoc'
  /knowledge/diagnose:
    post:
      tags: [知识库]
      summary: 智能诊断
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiagnoseRequest'
      responses:
        '200':
          description: 诊断结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnoseResponse'

  # ============ 报表 ============
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT认证令牌
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API密钥认证
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
    Device:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [s7, modbus]
        host:
          type: string
        port:
          type: integer
        status:
          type: string
          enum: [online, offline, error]
        last_seen:
          type: string
          format: date-time
        tag_count:
          type: integer
    DeviceTag:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        data_type:
          type: string
        unit:
          type: string
        description:
          type: string
    DeviceCreateRequest:
      type: object
      required: [name, type, host, port]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [s7, modbus]
        host:
          type: string
        port:
          type: integer
        rack:
          type: integer
        slot:
          type: integer
        tags:
          type: array
          items:
            $ref: '#/components/schemas/DeviceTag'
    DeviceUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/DeviceCreateRequest'

    # 告警相关
...
    KnowledgeDoc:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        similarity_score:
          type: number
    DiagnoseRequest:
      type: object
      required: [symptoms]
      properties:
        symptoms:
          type: string
          description: 症状描述
        device_id:
          type: string
        tags:
          type: object
          additionalProperties:
            type: number
    DiagnoseResponse:
      type: object
      properties:
        diagnosis_id:
          type: string
        root_cause:
          type: string
        confidence:
          type: number
        suggestions:
          type: array
          items:
            type: string
        spare_parts:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              quantity:
                type: integer
        references:
          type: array
          items:
            $ref: '#/components/schemas/KnowledgeDoc'

    # 报表相关
...
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...